optimized background service for independent component
